+++ app/jobs/cache/feed_job.rb
@@ -1,14 +1,6 @@
 # frozen_string_literal: true
 module Cache
   class FeedJob < ApplicationJob
-    def self.custom_feeds
-      {
-        custom: ::Feed.all.each_with_object({}) do |feed, hash|
-          hash[feed.slug.to_sym] = feed.url
-        end
-      }
-    end
-
     URLS = {
       blog: {
         all: 'http://blog.europeana.eu/feed/',
@@ -19,7 +11,7 @@ module Cache
       exhibitions: %i(de en).each_with_object({}) do |locale, hash|
         hash[locale] = (ENV['EXHIBITIONS_HOST'] || 'http://www.europeana.eu') + "/portal/#{locale}/exhibitions/feed.xml"
       end
-    }.merge custom_feeds
+    }
 
     queue_as :default
