+++ spec/fixtures/feeds.yml
@@ -1,4 +1,4 @@
 fashion_tumblr:
   name: Fashion Tumblr
-  slug: fashion_tumblr
+  slug: fashion-tumblr
   url: 'http://europeanafashion.tumblr.com/rss'
