+++ spec/models/feed_spec.rb
@@ -16,4 +16,20 @@ RSpec.describe Feed do
       expect(feed.to_param).to eq('art-tumblr')
     end
   end
+
+  describe '#url_in_domain?' do
+    context 'when URL is "http://www.europeana.eu/"' do
+      let(:url) { 'http://www.europeana.eu/' }
+
+      context 'domain is europeana.eu' do
+        subject { described_class.new(url: url).url_in_domain?('europeana.eu') }
+        it { is_expected.to be true }
+      end
+
+      context 'domain is europeana.com' do
+        subject { described_class.new(url: url).url_in_domain?('europeana.com') }
+        it { is_expected.to be false }
+      end
+    end
+  end
 end
